Adds the crunch that makes the cache updates work.
Now I just need to see if ti actually works, because I can't test things
on the train. I'm skeptical, but I hope I can get it to run.
Then I just need to make the json part (which is reasonably easy with
Cheshire and clj-http, and add the high-level interface. Whee!